вынес выподающий список валют в отдельный контрол
