+++ shared/components/modals/WithdrawModal/components/CurrencyList.scss
@@ -1,0 +1,123 @@
+.coin {
+  position: relative;
+  z-index: 1111111;
+}
+
+.amount {
+  position: absolute;
+  text-align: right;
+  right: 60px;
+  font-size: 13px;
+  @include media-mobile {
+    right: 50px;
+  }
+  .currency {
+    color: #1f2d48;
+  }
+  .usd {
+    color: #8e9aa3;
+  }
+}
+
+.customSelectListItem {
+  margin-top: -1px;
+  padding-left: 10px;
+  border: none;
+  &:hover {
+    background: #f3f3f3;
+  }
+}
+
+.customSelectArrow {
+  position: absolute;
+  right: 25px;
+  &:before {
+    content: '';
+    width: 9px;
+    height: 9px;
+    display: block;
+    border-top: 2px solid #828795;
+    border-right: 2px solid #828795;
+    -webkit-transform: rotate(135deg);
+    transform: rotate(135deg);
+  }
+}
+
+.customSelectArrow.active {
+  &:before {
+    -webkit-transform: rotate(-45deg);
+    transform: rotate(-45deg);
+  }
+}
+
+.customSelectListItem.disabled {
+  position: relative;
+  pointer-events: none;
+  &:before {
+    position: absolute;
+    content: '';
+    background: rgba(255, 255, 255, 0.8);
+    width: 100%;
+    height: 100%;
+    z-index: 1111;
+  }
+}
+
+.customSelectValue {
+  border: 1px solid #ccc;
+  height: 68px;
+  background: #fff;
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  div {
+    display: flex;
+    flex-direction: column;
+    margin-left: 10px;
+    a {
+      font-size: 13px;
+      color: #6144e5;
+      &:hover {
+        color: #6144e5;
+      }
+    }
+    .address {
+      color: #8e9aa3;
+      font-size: 12px;
+      @include media-mobile {
+        display: none;
+      }
+    }
+    .mobileAddress {
+      color: #8e9aa3;
+      font-size: 12px;
+      > div {
+        margin: 0;
+      }
+    }
+  }
+}
+
+.customSelectList {
+  margin-top: 10px;
+  position: absolute;
+  z-index: 1111111;
+  width: 100%;
+  max-height: 300px;
+  background: #fff;
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding: 5px 0px;
+  box-shadow: 0px 4px 16px rgba(0,0,0,0.08), 0px 1px 4px rgba(0,0,0,0.08);
+}
+
+@media all and (max-width: 480px) {
+  .customSelectValue {
+    cursor: pointer;
+    div {
+      .mobileAddress {
+        font-size: 8px;
+      }
+    }
+  }
+}
